Communication logic refactoring finished
Finished the refactoring described in the previous commit - moving the main
communication logic up from MasClient to MasCommunicationServices by means
of the template method pattern. MasCommunicationServices now handles
essentially all of the communication logic (including obtaining/setting the
session key returned by the server) except for the actual creating, reading,
writing, and closing of the TCP socket, which is done in MasClient.  For
"abstract methods", the 'abstract' gem/module is no longer used.  This
allows pre/post-conditions (via ruby_contract) to be defined for the
abstract methods and checked in the implementations (in MasClient).

Also - a bit of cleanup - remove old code, etc.  And added the
.gitignore file.