Tools for creating indicators/generators
cci, cci_signals, indicators, market_analyzers, parabolic-sar: "macl"
scripts to create indicators and tradable analyzers (indicators_persist
and generators_persist files) for the MAS server - to be used in tests.

make_indicators - bash script to feed the above files to macl, which
connects to the server and creates the ordered indicators/analyzers.