Finished 'request_analysis' with new TradableEvent class.
Added TradableEvent class and used it to finish the implementation of
'request_analysis' in MasCommunicationServices.  'request_analysis' creates
one TradableEvent for each line in the server's response and uses the event
id (3rd tab-separated field) to associate the new TradableEvent with the
correct TradableAnalyzer (and, by extension, the particular
tradable-event-generator on the server side that generated the event).  The
resulting array of "TradableEvent"s is found in the new attribute
'analysis_result'.  Completed 'test_analysis' in the test script to obtain
the generated events and assert that they are, indeed, "TradableEvent"s.

Also - in the test script, added a "destructor" routine via
define_finalizer that logs $client (global MasClient variable) out to
prevent leaving an open session, with an unused session-key, on the server
side.