Socket-communiction protocol change
Changes corresponding to a change in the server - alternate communication
protocol, allowing the socket to stay open after a
client-request/response pair - that is, the server (based on a
configuration) can behave as before, closing the socket after each
request/response (requiring the client to create a new TCPSocket before
each request/response) or it can be configured to keep the socket open.
If the server is configured for the original behavior, it will send an
OK_WILL_CLOSE status-code instead of an OK code, so that the client
will know when to close the connection.  (See mas_communication_protocol.rb
and mas_communication_services.rb.)  (Similar alternate codes will be needed
for the error-status values.)

A few more, probably minor, changes will be needed once the
implementation is finished on the server; and more testing (and probably
bug-fixing) and cleanup will be needed.