Param test changes, *_persist files
Added indicators_persist and generators_persist - binary
"indicators" and "analyzers" files, which were created by running
the following mas/macl scripts from <MASROOT>/build/config (where
<MASROOT> is the root directory for the files for the MAS
project): indicators, market_analyzers, cci, cci_signals. Since
the old generators_persist file is no longer compatible with the
MAS server, these files are now to be used for testing.

Added a couple parameter-related tests to test_mas_client.rb,
using the changed protocol that the parameters in the server
response are sorted on a unique parameter name - the client needs
to keep track of that order so that he can request the correct
index (e.g., index 3 for the 3rd param in the server response)
for a parameter change.