Some tuning of MasCommunicationServices/errors
MasCommunicationServices: MasServerError: new error type to
indicate error status returned from the server (as opposed to
system errors, such as socket connection or read errors).
'communication_failed' is now explicitly for system errors (such as
socket errors) and not for MAS server errors (i.e., not for
"MasServerError"s). 'execute_request' now raises an instance of the
new MasServerError class when server says not OK - i.e.:
       if not response_ok? then
         raise MasServerError.new(
           "Server returned error status: #{last_response}")
       end
Left a little $log.debug... lines in for debugging the time-out and
exception-related behavior - to be removed later.