chnaged test due to MAS end-date enforcement
ParameterTest.test_analyzer_parameters_modification2: One component of
this test started failing after commit
549617d12862b60d0ae611c9adcfee6673a9ce47 of the MAS server, which adds
an early implementation of enforcing end-dates for analysis. The
result is that the end dates specified in this test are actually used
by the server now, as opposed to before, when only the start date was
used and the analysis simply ended for a stock/tradable when the
last/latest record was encountered.  This difference caused different
results from previous test runs, which caused the one component to
"fail".  "Fixed" by changing the end date to the year 2038.