MAS change: addition of period-type for events
The client-server protocol has changed to provide a period-type
for event-processing to allow different period types to be used
(instead of an unchangeable, "preset", period-type, as before).
This means that a period-type must now be included in client
requests related to event processing (evgen_param_req,
event_param_set_req, and event_data_req from the server
communication protocol).

MasCommunicationServices: The period-type (or period-types [array],
in the case of request_analysis) has been added as an argument to:
'request_analysis_parameters', 'request_analysis_parameters_modification',
and 'request_anslysis' so that the period-type can be sent to the
server to conform to the protocol change.  Fixed a bug re. error
handling logic on line 324
    (if ! communication_failed && !  server_error then...).

Modified the tests to conform to this period-type protocol change.
Added some new tests (in parameter_test.rb) to specifically test
with different period types: daily, weekly, monthly.