Changes corresponding to TAT environment changes
The commit to the TAT project labeled:
"Enhancements to bin/runprod, rundev" required changes to 'startmas' and
test_setup.rb: startmas was moved from ./test to ./bin and now uses
the env. var. MAS_RUNDIR, set by .env in TAT, as the directory in which
to run the mas server.  test_setup.rb needed to change to find startmas in
bin instead of test directory.